# api/openapi.yaml
openapi: 3.1.0
info:
  title: WAM — WhatsApp Marketing API
  version: 0.1.0
  description: |
    WAM — WhatsApp Marketing. Source of truth for HTTP APIs.

    * UI: `GET /api-docs/` (Swagger UI). Spec: `GET /api-docs/openapi.yaml`
      (alias `GET /openapi.yaml`). Legacy `/swagger*` 301s to `/api-docs/`.
    * Dashboard: `/dashboard /connect /contacts /templates /campaigns /analytics`
      (htmx, no React). Labels are embedded in Contacts (`/contacts#labels`), not a separate nav.
    * Connection pairing (QR/phone/logout/status), number check,
      one-off send, contacts/labels CRUD + import, templates CRUD + duplicate,
      campaigns + worker, per-campaign funnel, account overview, CSV export.
    * Persistence: SQLite (`WAM_DB_PATH`), shared with whatsmeow sqlstore.
      Auth: single admin (`WAM_ADMIN_PASSWORD`), session cookie.
    * SaaS path (Postgres `WAM_DATABASE_URL` + `WAM_APP_DATABASE_URL`):
      DB users + RBAC (`admin`/`user`, permission checks per handler),
      per-request org transactions (fail-closed RLS), per-account grants
      (`/api/v1/grants`, enforced in multi-account), signup/verify/reset,
      invites, multi-org sessions with switcher.
  contact:
    name: WAM
  license:
    name: UNLICENSED
    identifier: UNLICENSED
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://wam.example.com
    description: Production
...
paths:
...
  /login:
    post:
      tags: [Auth]
      operationId: login
      summary: Sign in (requires verified email). JSON API or web form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: OK (sets session cookie) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Email unverified }
...
  /api/v1/members:
    get:
      tags: [Members]
      operationId: listMembers
      summary: Members + pending invites (admin)
      responses:
        '200': { description: OK }
        '403': { description: Missing members:manage }
  /api/v1/members/invite:
    post:
      tags: [Members]
      operationId: inviteMember
      summary: Invite by email + role (admin, mails link)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                role: { type: string, enum: [admin, user], default: user }
      responses:
        '201': { description: Invited }
        '403': { description: Missing members:manage }
  /api/v1/members/{user_id}:
    patch:
      tags: [Members]
      operationId: setMemberRole
      summary: Change role (last-admin guard → 409)
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [admin, user] }
      responses:
        '200': { description: OK }
        '409': { description: Last admin }
    delete:
      tags: [Members]
      operationId: removeMember
      summary: Remove member + revoke org sessions (last-admin guard → 409)
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Removed }
        '409': { description: Last admin }
  /api/v1/grants:
    get:
      tags: [Members]
      operationId: listGrants
      summary: Account grants for a member (?user_id=, admin)
      parameters:
        - name: user_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    post:
      tags: [Members]
      operationId: setGrant
      summary: Grant member access to an account (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, account_id, role]
              properties:
                user_id: { type: string }
                account_id: { type: string }
                role: { type: string, enum: [member, viewer] }
      responses:
        '200': { description: OK }
    delete:
      tags: [Members]
      operationId: removeGrant
      summary: Revoke account access (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, account_id]
              properties:
                user_id: { type: string }
                account_id: { type: string }
      responses:
        '204': { description: Revoked }
components:
...
  schemas:
...
    Problem:
      title: RFC 9457 problem+json
      type: object
      required: [title, status]
      properties:
        type: { type: string, format: uri, default: 'about:blank' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
  responses:
...
    Unauthorized:
      description: Login required
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }